convert pad_pre and pad_post to int

+++ tomviz/python/Recon_DFT.py
@@ -33,9 +33,9 @@ def dfm3(input, angles, Npad):
     input = np.double(input)
     (Nx, Ny, Nproj) = input.shape
     angles = np.double(angles)
-    pad_pre = np.ceil(
-        (Npad - Ny) / 2.0)
-    pad_post = np.floor((Npad - Ny) / 2.0)
+    pad_pre = int(np.ceil(
+        (Npad - Ny) / 2.0) )
+    pad_post = int(np.floor((Npad - Ny) / 2.0))
 
     # Initialization
     Nz = Ny
